// ms-product/src/models/requests/search-product.request.ts
export class SearchProductBy {
    id: number;
    name: string;
    price: number;
    quantity: number;
    status: number; 
    createdBy: number; 
}

export class searchProductReportBy {
    id: number;
    name: string;
    price: number;
    quantity: number;
    status: number; 
    createdBy: number; 
    take: number;
    skip: number;
}